+++ nflow-engine/src/main/java/io/nflow/engine/internal/workflow/ObjectStringMapper.java
@@ -4,28 +4,26 @@ import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Type;
 
-import io.nflow.engine.config.EngineConfiguration.EngineObjectMapperSupplier;
-import jakarta.inject.Inject;
-
 import org.springframework.stereotype.Component;
 
-import tools.jackson.core.JacksonException;
-import tools.jackson.databind.JavaType;
-import tools.jackson.databind.ObjectMapper;
-
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import io.nflow.engine.config.EngineConfiguration.EngineJsonMapperSupplier;
 import io.nflow.engine.config.NFlow;
 import io.nflow.engine.internal.workflow.WorkflowStateMethod.StateParameter;
 import io.nflow.engine.workflow.definition.Mutable;
 import io.nflow.engine.workflow.definition.StateExecution;
+import jakarta.inject.Inject;
+import tools.jackson.core.JacksonException;
+import tools.jackson.databind.JavaType;
+import tools.jackson.databind.json.JsonMapper;
 
 @Component
 public class ObjectStringMapper {
-  private final ObjectMapper mapper;
+  private final JsonMapper mapper;
 
   @Inject
-  public ObjectStringMapper(@NFlow EngineObjectMapperSupplier nflowObjectMapper) {
-    this.mapper = nflowObjectMapper.get();
+  public ObjectStringMapper(@NFlow EngineJsonMapperSupplier nflowJsonMapper) {
+    this.mapper = nflowJsonMapper.get();
   }
 
   @SuppressWarnings("unchecked")