+++ nflow-rest-api-common/src/test/java/io/nflow/rest/v1/converter/ListWorkflowInstanceConverterTest.java
@@ -37,13 +37,13 @@ import io.nflow.rest.v1.msg.ListWorkflowInstanceResponse;
 import tools.jackson.core.JacksonException;
 import tools.jackson.core.exc.StreamReadException;
 import tools.jackson.databind.JsonNode;
-import tools.jackson.databind.ObjectMapper;
+import tools.jackson.databind.json.JsonMapper;
 import tools.jackson.databind.node.StringNode;
 
 @ExtendWith(MockitoExtension.class)
 public class ListWorkflowInstanceConverterTest {
   @Mock
-  private ObjectMapper nflowObjectMapper;
+  private JsonMapper nflowJsonMapper;
   @InjectMocks
   private final ListWorkflowInstanceConverter converter = new ListWorkflowInstanceConverter();
 
@@ -65,8 +65,8 @@ public class ListWorkflowInstanceConverterTest {
 
     JsonNode node1 = mock(JsonNode.class);
     JsonNode nodeQuux = mock(JsonNode.class);
-    when(nflowObjectMapper.readTree("1")).thenReturn(node1);
-    when(nflowObjectMapper.readTree("quux")).thenReturn(nodeQuux);
+    when(nflowJsonMapper.readTree("1")).thenReturn(node1);
+    when(nflowJsonMapper.readTree("quux")).thenReturn(nodeQuux);
 
     Map<String, Object> expectedStateVariables = new LinkedHashMap<>();
     expectedStateVariables.put("foo", node1);
@@ -74,8 +74,8 @@ public class ListWorkflowInstanceConverterTest {
 
     ListWorkflowInstanceResponse resp = converter.convert(i, EnumSet.of(actions, currentStateVariables), false);
 
-    verify(nflowObjectMapper).readTree("1");
-    verify(nflowObjectMapper).readTree("quux");
+    verify(nflowJsonMapper).readTree("1");
+    verify(nflowJsonMapper).readTree("quux");
     assertThat(resp.id, is(i.id));
     assertThat(resp.stateVariables, is(expectedStateVariables));
     assertThat(resp.status, is(i.status.name()));
@@ -115,8 +115,8 @@ public class ListWorkflowInstanceConverterTest {
 
     JsonNode node1 = mock(JsonNode.class);
     JsonNode nodeQuux = mock(JsonNode.class);
-    when(nflowObjectMapper.readTree("1")).thenReturn(node1);
-    when(nflowObjectMapper.readTree("quux")).thenReturn(nodeQuux);
+    when(nflowJsonMapper.readTree("1")).thenReturn(node1);
+    when(nflowJsonMapper.readTree("quux")).thenReturn(nodeQuux);
 
     Map<String, Object> expectedStateVariables = new LinkedHashMap<>();
     expectedStateVariables.put("foo", node1);
@@ -124,8 +124,8 @@ public class ListWorkflowInstanceConverterTest {
 
     ListWorkflowInstanceResponse resp = converter.convert(i, EnumSet.of(actions, actionStateVariables), true);
 
-    verify(nflowObjectMapper).readTree("1");
-    verify(nflowObjectMapper).readTree("quux");
+    verify(nflowJsonMapper).readTree("1");
+    verify(nflowJsonMapper).readTree("quux");
     assertThat(resp.id, is(i.id));
     assertThat(resp.status, is(i.status.name()));
     assertThat(resp.type, is(i.type));
@@ -252,13 +252,13 @@ public class ListWorkflowInstanceConverterTest {
         .setStatus(inProgress).setType("dummy").setBusinessKey("businessKey").setExternalId("externalId").setState("cState")
         .setStateText("cState desc").setNextActivation(now).setActions(asList(a)).setStateVariables(stateVariables).build();
 
-    when(nflowObjectMapper.readTree(value1)).thenThrow(new StreamReadException("bad data"));
-    when(nflowObjectMapper.readTree(value2)).thenThrow(new StreamReadException("bad data"));
+    when(nflowJsonMapper.readTree(value1)).thenThrow(new StreamReadException("bad data"));
+    when(nflowJsonMapper.readTree(value2)).thenThrow(new StreamReadException("bad data"));
 
     ListWorkflowInstanceResponse resp = converter.convert(i, EnumSet.of(currentStateVariables), false);
 
-    verify(nflowObjectMapper).readTree(value1);
-    verify(nflowObjectMapper).readTree(value2);
+    verify(nflowJsonMapper).readTree(value1);
+    verify(nflowJsonMapper).readTree(value2);
     Map<String, Object> expectedStateVariables = new LinkedHashMap<>();
     expectedStateVariables.put("foo", new StringNode(value1));
     expectedStateVariables.put("bar", new StringNode(value2));