+++ nflow-rest-api-jax-rs/src/test/java/io/nflow/rest/v1/config/jaxrs/RestConfigurationTest.java
@@ -1,17 +1,16 @@
 package io.nflow.rest.v1.config.jaxrs;
 
-import static tools.jackson.databind.cfg.DateTimeFeature.WRITE_DATES_AS_TIMESTAMPS;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
+import static tools.jackson.databind.cfg.DateTimeFeature.WRITE_DATES_AS_TIMESTAMPS;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import tools.jackson.databind.json.JsonMapper;
-
 import io.nflow.rest.config.RestConfiguration;
+import tools.jackson.databind.json.JsonMapper;
 
 @ExtendWith(MockitoExtension.class)
 public class RestConfigurationTest {
@@ -24,8 +23,8 @@ public class RestConfigurationTest {
   }
 
   @Test
-  public void nflowRestObjectMapperInstantiated() {
-    JsonMapper restMapper = configuration.nflowRestObjectMapper(JsonMapper::new);
-    assertThat(restMapper.serializationConfig().isEnabled(WRITE_DATES_AS_TIMESTAMPS), is(false));
+  public void nflowRestJsonMapperInstantiated() {
+    JsonMapper mapper = configuration.nflowRestJsonMapper(JsonMapper::new);
+    assertThat(mapper.serializationConfig().isEnabled(WRITE_DATES_AS_TIMESTAMPS), is(false));
   }
 }