+++ nflow-tests/src/test/java/io/nflow/tests/MaintenanceTest.java
@@ -150,7 +150,7 @@ public class MaintenanceTest extends AbstractNflowTest {
   private long createWorkflow() {
     CreateWorkflowInstanceRequest req = new CreateWorkflowInstanceRequest();
     req.type = FIBONACCI_TYPE;
-    req.stateVariables.put(VAR_REQUEST_DATA, nflowObjectMapper().valueToTree(new FibonacciWorkflow.FiboData(3)));
+    req.stateVariables.put(VAR_REQUEST_DATA, nflowJsonMapper().valueToTree(new FibonacciWorkflow.FiboData(3)));
     CreateWorkflowInstanceResponse resp = createWorkflowInstance(req);
     assertThat(resp.id, notNullValue());
     return resp.id;